requirements.txt
flake8==3.9.2
flake8-docstrings==1.6.0
pytest==6.2.5
python-dotenv==0.19.0
python-telegram-bot==13.7
requests==2.26.0

## для правильного оформления

Django==2.2.19
flake8==3.9.2
flake8-broken-line==0.3.0
flake8-isort==4.0.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
isort==5.9.3
mccabe==0.6.1
pep8-naming==0.12.1
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1